Create solution.cpp
Convert a Roman Numeral to Int equivalent. Easiest way is probably to just create a map. That way, the key is the roman numeral character and the value returned is the integer equivalent. You just sum up the values until you reach a case of I,X, or C coming before a larger roman numeral. You just subtract I, X, or C from the larger value and skip the next iteration of the loop so it doesn't add the larger value again.